add env vars to leave unused resources in file
This change can be used to bisect a resource that is being too
aggresively removed by gfxrecon-optimize.

Can set START_KEEPING and END_KEEPING to force gfxrecon-optimize to
leave in resources that are in that range inclusive.

If range length is one (START == END), print the ID of the resource
forced in.