dx12 dump: Split command lists
Split the duplicated command list into three command lists.
The first is the begining to the before target drawcall.
The second is the binding, setting and the target drawcall.
The third is the binding, setting and the after target drawcall.

It helps to add new command list to before and after target drawcall.